* quickrestart system replaced with a function that changes video settings in mainmenu code (it's in mainmenu in order to correctly free OpenGL resources before  changing settings) * no more --restore command line option * --quickrestart command line option replaced with --quickstart, it does the same thing except you get in main menu and not in options * removed scripts that used to start ta3d and 3dmeditor, since we now use packages on Linux, they're not needed any more and changing video settings is done in the code itself (note that it's faster now :))
this fixes #44

+++ src/gfx/gui/area.cpp
@@ -413,7 +413,10 @@ namespace TA3D
         vec_wnd.clear();			// Empty the window vector
         vec_z_order.clear();		// No more windows at end
 
-        gfx->destroy_texture(background); // Destroy the texture
+        if (background == gfx->glfond)      // Don't remove the background texture
+            background = 0;
+        else
+            gfx->destroy_texture(background); // Destroy the texture
         if (skin)					// Destroy the skin
             delete skin;
     }
